+++ src/providers/mod.rs
@@ -4,6 +4,8 @@
 
 use super::{Greeting, Scrape, ScrapeError, ScrapeResult, Url};
 
+mod buongiornissimo_caffe;
 mod ilmondodigrazia;
 
+pub use buongiornissimo_caffe::BuongiornissimoCaffe;
 pub use ilmondodigrazia::IlMondoDiGrazia;