+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.2.0](#020)
   - [0.1.0](#010)
 
 ---
+
+## 0.2.0
+
+Released on 12/09/2022
+
+- Added new greetings
+  - san valentino
+  - festa della donna
+  - domenica delle palme
+  - santo stefano
+- added `BuongiornissimoCaffe` provider from <https://www.buongiornissimocaffe.it>
 
 ## 0.1.0